Don't update inverse relation if it hasn't changed
Prevents flickering

+++ src/model.coffee
@@ -69,8 +69,9 @@ class Model
           newCollection = model?[options.inverseOf] or new Collection()
           oldCollection = previous?[options.inverseOf] or new Collection()
 
-          oldCollection.delete(this)
-          newCollection.push(this) unless this in newCollection
+          unless oldCollection is newCollection
+            oldCollection.delete(this)
+            newCollection.push(this) unless this in newCollection
 
     @property name, propOptions
     @property name + 'Id',
